requirements.txt
Django==4.0.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
django-filter==21.1
celery==5.2.7
django-crontab==0.7.1